LPS-83488 Move logic over to ResourcePermissionLocalService and deligate from ResourceLocalService. Rename methods accordingly.
